Generic GET methods for Controllers of linus and vla DB

+++ VLA-Backend/src/main/java/de/vlaorgatu/vlabackend/repositories/linusdb/LinusExperimentRepository.java
@@ -5,5 +5,5 @@ import de.vlaorgatu.vlabackend.entities.linusdb.LinusExperiment;
 /**
  * Repository for accessing and managing {@link LinusExperiment} entities.
  */
-public interface LinusExperimentRepository extends ReadOnlyRepository<LinusExperiment, Long> {
+public interface LinusExperimentRepository extends ReadOnlyRepository<LinusExperiment, Integer> {
 }
