+++ VLA-Backend/src/main/java/de/vlaorgatu/vlabackend/repositories/linusdb/LinusUserRepository.java
@@ -5,5 +5,5 @@ import de.vlaorgatu.vlabackend.entities.linusdb.LinusUser;
 /**
  * Repository for accessing and managing {@link LinusUser} entities.
  */
-public interface LinusUserRepository extends ReadOnlyRepository<LinusUser, Long> {
+public interface LinusUserRepository extends ReadOnlyRepository<LinusUser, Integer> {
 }
